+++ android/src/main/kotlin/com/tyx/flutter_matter/BluetoothManager.kt
@@ -1,4 +1,4 @@
-package com.zengge.flutter_matter
+package com.tyx.flutter_matter
 
 import android.annotation.SuppressLint
 import android.bluetooth.BluetoothAdapter