+++ android/src/test/kotlin/com/tyx/flutter_matter/FlutterMatterPluginTest.kt
@@ -1,4 +1,4 @@
-package com.zengge.flutter_matter
+package com.tyx.flutter_matter
 
 import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel