+++ example/lib/onboarding.dart
@@ -1,9 +1,6 @@
 import 'package:flutter/cupertino.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/widgets.dart';
 import 'package:flutter_matter/flutter_matter.dart';
-import 'package:flutter_matter_example/device_provisioning.dart';
 import 'package:flutter_matter_example/wifi_input.dart';
 import 'package:fluttertoast/fluttertoast.dart';
 import 'package:mobile_scanner/mobile_scanner.dart';